Restructure folders into controllers, services, interfaces and validators
