+++ src/app.ts
@@ -1,8 +1,8 @@
+import { router } from "@src/routes"
 import cookieParser from "cookie-parser"
 import cors from "cors"
 import express, { Application } from "express"
 import helmet from "helmet"
-import { router } from "./routes"
 /**
  * Todo: Use express-session instead of cookie-parser?
  * https://www.npmjs.com/package/express-session
@@ -18,5 +18,3 @@ app.use(cors())
 app.use(helmet())
 app.use(cookieParser())
 app.use(baseUrl, router)
-
-
