+++ src/controllers/reservations/index.ts
@@ -1,5 +1,5 @@
+import { newReservationController } from "@src/controllers/reservations/NewReservation/newReservationController"
 import { Router } from "express"
-import { newReservationController } from "./controllers/NewReservation/newReservationController"
 
 export const useReservationControllers = (router: Router) => {
   newReservationController(router)
