+++ src/controllers/users/index.ts
@@ -1,5 +1,5 @@
+import { LoginController } from "@src/controllers/users/LoginController"
 import { Router } from "express"
-import { LoginController } from "./controller/LoginController"
 
 export const useUserControllers = (router: Router) => {
   LoginController(router)
