+++ src/validators/users/validateIsAuthenticated.ts
